// 23-sistemas-monoliticos/code-store/src/modules/store-catalog/facade/store-catalog.facade.interface.ts
export interface FindStoreCatalogFacadeInputDto {
    id: string;
  }
  
export interface FindStoreCatalogFacadeOutputDto {
  id: string;
  name: string;
  description: string;
  salesPrice: number;
}
  
export interface FindAllStoreCatalogFacadeOutputDto {
  products: {
    id: string;
    name: string;
    description: string;
    salesPrice: number;
  }[];
}
  
export default interface IStoreCatalogFacade {

  find(id: FindStoreCatalogFacadeInputDto): Promise<FindStoreCatalogFacadeOutputDto>;
  findAll(): Promise<FindAllStoreCatalogFacadeOutputDto>;
}